fix 予めcontentを@HPか空に変換しておくようにした
